refactor: 객체를 포장하는 책임을 Controller로 변경, 컨트롤러 클래스 이름 수정
